+++ core-ng-search/src/main/java/core/framework/search/ElasticSearch.java
@@ -4,9 +4,10 @@ package core.framework.search;
  * @author neo
  */
 public interface ElasticSearch {
-    void createIndex(String index, String source);
+    // create index or update mappings
+    void putIndex(String index, String source);
 
-    void createIndexTemplate(String name, String source);
+    void putIndexTemplate(String name, String source);
 
     void closeIndex(String index);